Sample of how this JFR code could look
Adds a simple JFR event and util class, and uses it from two places in
the compiler, plus updates SpeedTracerLogger to write all logs to it.

The output is pretty terrible (turns out we almost definitely want
specialized Event types per use case, as well as carefully structured
categories and so on), but you definitely can see the structure of the
compile if you work at it.

Edited samples to enable jfr and speedtracer (with one worker).

Fixed a confusingly labeled event for DuplicateClinitRemover, and edited
a few other keys as well.