Deprecate and disable CollectorLegacy, simplify startup (#10185)
CollectorLegacy shouldn't be needed in any modern browser, though the
Error.stack property is still non-standard. This patch leaves it in, so
that it is at least possible for projects to restore it if they need it,
but won't add an extra step to startup for most GWT applications.

With only a single implementation, the static final field isn't needed
and likewise doesn't need to be eagerly initialized. This will remove
clinits for both Impl and StackTraceCreator everywhere they are used,
and by default should still compile out the collector instance into just
a set of static functions.