Enhance AstDumper for debugging and testing (#10177)
Enhances the AstDumper utility class with the aim of bringing back some
of the functionality previously removed in
64f9b00adb7e29f9247b95bd0706823de1472e6b, where individual methods could
be traced from source Java all the way to JS output. This is a small
step in that direction that nevertheless provides more insight into what
the compiler is doing, through a series of changes:
* Improve printed Java output - classes should list their name, enums
and records should be labeled as such, etc. AST traversal is made more
visitor-like to make filtering easier.
* JS AST should also be printed, rather than ending with the Java
optimization loops, allowing insight into JS passes.
* A filter can be applied to the ast at any stage, resulting in only
printing code from the specified files or lines, to more easily follow
specific changes through a program.

More improvement remains - multiple permutations may end up mixed in the
same file, different compiler passes are not individually listed or
labeled, arguments are re-parsed every time they are used, but impact to
the API is minimal for now.