Update Jetty to 9.4.58.v20250814 (#10203)
This pull requests updates the jetty libraries due to a security issue.
I know that gwt-dev is not for production code however explain this to
the security guys ;-)

There is a [related PR](https://github.com/gwtproject/tools/pull/38) in
the gwt-tools project as well.

**Please note,** that the checks are failing since the libs are missing
from the gwt-tools project. I think there is no way to link both
project's here.

---------

Co-authored-by: Zbynek Konecny <zbynek1729@gmail.com>
Co-authored-by: Colin Alworth <colin@vertispan.com>